ISSUE-186 - Moved "clinvar_identifier_parsing.py" to the package And amended its imports within the repo
